Make mpi_message_limit an unsigned int, switch units to KB

+++ src/scheduler.h
@@ -105,7 +105,7 @@ struct scheduler {
     
   /* Maximum size of task messages, in bytes, to sent using non-buffered
    * MPI. */
-  int mpi_message_limit;
+  size_t mpi_message_limit;
 
   /* 'Pointer' to the seed for the random number generator */
   pthread_key_t local_seed_pointer;
